apple cleanup
Added more erodes with smaller structure elements to take less of the
apple away but more of the background.

Almost doubled the closing structure element size to get all holes
removed. Because all of the background pixels are removed with the
erode, this is no problem.

Tripled dilation structure element size to get closer to the original
size of the apple.